+++ packages/netlify/src/env.d.ts
@@ -1,1 +1,11 @@
 /// <reference types="astro/client" />
+
+import type { Context } from '@netlify/functions';
+
+declare namespace App {
+	interface Locals {
+		netlify: {
+			context: Context;
+		};
+	}
+}
